Moved Broker and Proxy to the package data
